bx操作记录保存

+++ src/main/java/com/zhdtedu/system/config/CustomRequestMappingHandlerMapping.java
@@ -25,6 +25,6 @@ public class CustomRequestMappingHandlerMapping extends RequestMappingHandlerMap
     }
 
     private RequestCondition<ApiVesrsionCondition> createCondition(ApiVersion apiVersion) {
-        return apiVersion == null ? null : new ApiVesrsionCondition(apiVersion.value());
+        return apiVersion == null ? null : new ApiVesrsionCondition((int) apiVersion.value());
     }
 }
